Some slight fixes and balancing to the economy

+++ src/main/java/dev/darealturtywurty/superturtybot/core/util/StringUtils.java
@@ -288,7 +288,7 @@ public final class StringUtils {
         }
     }
 
-    public static String formatCurrency(String currencySymbol, int amount) {
+    public static String formatCurrency(String currencySymbol, long amount) {
         NumberFormat currencyFormatter = NumberFormat.getCurrencyInstance(Locale.US);
         String formattedAmount = currencyFormatter.format(amount);
